requirements.txt
asgiref==3.8.1
boto3==1.38.42
botocore==1.38.42
Django==5.2.3
django-storages==1.14.6
djangorestframework==3.16.0
jmespath==1.0.1
pillow==11.2.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
